feat(workspace): clean all packages
