+++ packages/api/index.ts
@@ -1,8 +1,7 @@
-import { Web3 } from "./src/web3";
 import { API, ExResult } from "./src/api";
-import { autoAPI, autoWeb3 } from "./src/auto";
+import { autoAPI } from "./src/auto";
 import { ShadowAPI } from "./src/shadow";
 
 export {
-    API, autoAPI, autoWeb3, ExResult, ShadowAPI, Web3,
+    API, autoAPI, ExResult, ShadowAPI,
 }
