+++ packages/api/package.json
@@ -1,7 +1,7 @@
 {
   "private": false,
   "name": "@darwinia/api",
-  "version": "0.1.21",
+  "version": "0.1.22",
   "description": "High-level darwinia.js api",
   "homepage": "https://github.com/darwinia-network/darwinia.js",
   "repository": {
@@ -11,16 +11,13 @@
   "author": "darwinia-network <hello@darwinia.network>",
   "license": "GPL-3.0",
   "main": "lib/index.js",
-  "files": [
-    "lib/**/*"
-  ],
+  "files": ["lib/**/*"],
   "dependencies": {
-    "@darwinia/util": "^0.1.21",
+    "@darwinia/util": "^0.1.22",
     "@polkadot/api": "^1.10.1",
-    "@polkadot/keyring": "2.9.1",
-    "@polkadot/util-crypto": "2.9.1",
-    "axios": "^0.19.2",
-    "web3": "^1.2.6"
+    "@polkadot/keyring": "^2.9.1",
+    "@polkadot/util-crypto": "^2.9.1",
+    "axios": "^0.19.2"
   },
   "scripts": {
     "build": "tsc --strict",
