+++ packages/dj/package.json
@@ -1,7 +1,7 @@
 {
   "private": false,
   "name": "@darwinia/dj",
-  "version": "0.1.21",
+  "version": "0.1.22",
   "description": "High-level darwinia.js api",
   "homepage": "https://github.com/darwinia-network/darwinia.js",
   "repository": {
@@ -18,17 +18,10 @@
   "dependencies": {
     "@darwinia/api": "^0.1.21",
     "@darwinia/util": "^0.1.21",
-    "apollo-server": "^2.12.0",
-    "chalk": "^4.0.0",
-    "graphql-import-node": "^0.0.4",
-    "jayson": "^3.3.0",
-    "js-yaml": "^3.13.1",
-    "knex": "^0.20.14",
-    "sqlite3": "^4.1.1",
     "yargs": "^15.3.1"
   },
   "scripts": {
-    "build": "tsc --strict && cp src/static/schema.gql lib/src/static/schema.gql",
+    "build": "tsc --strict",
     "lint": "tsc --noEmit --strict && tslint --project ./tsconfig.json"
   },
   "gitHead": "0999fb93cce8f12e1599a2db882148798dcf0c33"
