--- packages/dj/src/static/schema.gql
@@ -1,7 +1,0 @@
-type Query {
-  "Let's talk!"
-  answer(
-    ident: String!,
-    message: String
-  ): String
-}
