+++ packages/util/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@darwinia/util",
-  "version": "0.1.21",
+  "version": "0.1.22",
   "description": "Util for developing darwinia javascript library.",
   "homepage": "https://github.com/darwinia-network/darwinia.js",
   "repository": {
@@ -12,16 +12,15 @@
   "main": "lib/index.js",
   "files": ["lib/**/*"],
   "dependencies": {
-    "@polkadot/util": "2.8.1",
+    "@polkadot/util": "^2.9.1",
     "chalk": "^4.0.0",
-    "ethereumjs-util": "^6.2.0",
+    "ethereumjs-util": "^7.0.0",
     "got": "^10.7.0",
     "javascript-stringify": "^2.0.1",
-    "progress": "^2.0.3",
-    "web3-utils": "^1.2.6"
+    "progress": "^2.0.3"
   },
   "scripts": {
-    "build": "tsc --strict && cp src/static/grammer.yml lib/src/static/grammer.yml",
+    "build": "tsc --strict",
     "lint": "tsc --noEmit --strict && tslint --project ./tsconfig.json"
   }
 }
