+++ packages/util/src/static/config.json
@@ -1,15 +1,4 @@
 {
-  "eth": {
-    "api": "",
-    "secret": ""
-  },
-  "grammer": {
-    "faucet": {
-      "supply": 400,
-      "amount": 1000000000000,
-      "interval": 24
-    }
-  },
   "node": "ws://0.0.0.0:9944",
   "seed": "//Alice",
   "shadow": "https://shadow.darwinia.network"
